+++ resources/translations/de_DE.ts
@@ -5095,17 +5095,17 @@ Tippe auf den QR-Code, um zu Vergrößern.</translation>
 <context>
     <name>uc::core::Api</name>
     <message>
-        <location filename="../../src/core/core.cpp" line="1362"/>
+        <location filename="../../src/core/core.cpp" line="1368"/>
         <source>Connection error</source>
         <translation>Verbindungsfehler</translation>
     </message>
     <message>
-        <location filename="../../src/core/core.cpp" line="1363"/>
+        <location filename="../../src/core/core.cpp" line="1369"/>
         <source>There was an error connecting to the core. If the issue persists, restart the remote.</source>
         <translation>Beim Verbinden auf den Core Service ist ein Fehler aufgetreten. Wenn das Problem weiterhin besteht, starte die Fernbedienung neu.</translation>
     </message>
     <message>
-        <location filename="../../src/core/core.cpp" line="1822"/>
+        <location filename="../../src/core/core.cpp" line="1828"/>
         <source>Authentication to core failed</source>
         <translation>Authentifizierung mit Core Service fehlgeschlagen</translation>
     </message>